Updated build.xml.
To simplify the build process the build.xml has been modified to
define default build properties that match Fedora platform. This
way the build can be done simply as follows:

 $ ant

For other platform the build properties can be defined in a file,
for example:

 $ ant -propertyfile rhel.properties

A new build target has also been added to install the binaries
directly on the system without building an RPM file.