Use get-current-version action for COPR builds
Currently Packit auotmagically overwrites the spec version with a
version derived from the latest tag when doing PR/commit COPR
builds. This causes problems as the tag can contain the alpha/beta
phase in addition to the version.

We don't need to do it for the propose_downstream job as the spec
is synced from the upstream repo so it will always provide the
correct version.